refactor wptfull into own class
