+++ main/src/main/java/cgeo/geocaching/utils/gpx/GPXUtils.java
@@ -1,6 +1,9 @@
 package cgeo.geocaching.utils.gpx;
 
+import cgeo.geocaching.location.Geopoint;
 import cgeo.geocaching.utils.Log;
+import cgeo.geocaching.utils.xml.XmlNode;
+import cgeo.geocaching.utils.xml.XmlUtils;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
@@ -10,13 +13,16 @@ import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.LinkedHashMap;
+import java.util.List;
 import java.util.Map;
+import java.util.Set;
 import java.util.zip.ZipEntry;
 
 import org.apache.commons.compress.archivers.zip.ZipArchiveInputStream;
 import org.apache.commons.io.IOUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.Strings;
+import org.xmlpull.v1.XmlPullParser;
 import org.xmlpull.v1.XmlPullParserException;
 
 /**
@@ -95,5 +101,79 @@ public final class GPXUtils {
         }
         return result;
     }
+
+    @Nullable
+    public static String readLinkUrl(@Nullable final XmlNode link) {
+        return StringUtils.defaultIfBlank(gpxNodeAttrValue(link, "href", null), gpxNodeChildText(link, "href", null));
+    }
+
+    @Nullable
+    public static String readWaypointUrl(final XmlNode node) {
+        return StringUtils.defaultIfBlank(node.getChildValue("url"), readLinkUrl(node.getChild("link")));
+    }
+
+    // ---------------------------------------------------------------------------------------------------
+    // date parsing: tolerant of the handful of date formats found in real-world GPX files
+    // ---------------------------------------------------------------------------------------------------
+
+    // ---------------------------------------------------------------------------------------------------
+    // XmlNode helpers: namespace-tolerant lookup by local name (D-namespace tolerance, point 34)
+    // ---------------------------------------------------------------------------------------------------
+
+    /**
+     * For GPX 1.1 files, extension fields live inside a {@code <extensions>} child; for GPX 1.0 files, they are
+     * direct children of the {@code wpt}/{@code rtept}/{@code trkpt} element itself. .
+     */
+    @Nullable
+    public static XmlNode extensionsBase(final XmlNode wptNode) {
+        final XmlNode extensions = gpxNodeChild(wptNode, "extensions", null);
+        return extensions != null ? extensions : wptNode;
+    }
+
+    @Nullable
+    public static XmlNode gpxNodeChild(final XmlNode node, final String name, final Set<String> namespaces) {
+        return XmlNode.getChild(node, name, namespaces);
+    }
+
+    @Nullable
+    public static String gpxNodeChildText(final XmlNode node, final String name, final Set<String> namespaces) {
+        final XmlNode c = gpxNodeChild(node, name, namespaces);
+        // A present, empty element must be able to clear an earlier fallback value.
+        return c == null ? null : StringUtils.defaultString(c.getValue());
+    }
+
+    /** Like {@link #gpxNodeChild}, but returns ALL matching children (namespace-tolerant, by local name), not just the first. */
+    @Nullable
+    public static List<XmlNode> gpxNodeChildren(final XmlNode node, final String name) {
+        return node == null ? null : node.getChildrenAsList(name);
+    }
+
+    @Nullable
+    public static String gpxNodeAttrValue(final XmlNode node, final String attributeName, final Set<String> namespaces) {
+        return gpxNodeChildText(node, XmlNode.ATTRIBUTE_PRAEFIX + attributeName, namespaces);
+    }
+
+    @Nullable
+    public static Geopoint gpxNodeReadLatLon(final XmlNode node) {
+        final String lat = gpxNodeAttrValue(node, "lat", null);
+        final String lon = gpxNodeAttrValue(node, "lon", null);
+        return XmlUtils.parseGeopoint(lat, lon, false);
+    }
+
+    @Nullable
+    public static String attr(final XmlPullParser parser, final String name) {
+        for (int i = 0; i < parser.getAttributeCount(); i++) {
+            if (name.equals(localName(parser.getAttributeName(i)))) {
+                return parser.getAttributeValue(i);
+            }
+        }
+        return null;
+    }
+
+    @Nullable
+    public static String localName(final String raw) {
+        return XmlUtils.getLocalName(raw);
+    }
+
 }
